commit
commit

+++ vam/14uclncuansodautien.cpp
@@ -3,7 +3,7 @@ using namespace std;
 using ll = long long;
 
 
-int lcm(int a, int b){
+ll lcm(ll a, ll b){
     return a / __gcd(a,b) * b;
 }
 
@@ -11,10 +11,10 @@ int main(){
     int t;
     cin >> t;
     while(t--){
-    int n;
+    ll n;
     cin >> n;
-    int res = 1;
-    for (int i = 1; i <= n; i++){
+    ll res = 1;
+    for (ll i = 1; i <= n; i++){
         res = lcm(i, res);
     }
     cout << res << endl;
